Assorted Arduino Project sketches
This section is for Arduino projects that i built. Many of this project are part of Design project from different engineering student that contracted me to build the entire project or write the code for their projects. I am adding this to my github repository as Legacy code for student and hobbyist that are willing to study and work hard for their project.